Code Refactoring

+++ data/src/main/kotlin/com/taetae98/diary/data/datasource/ExceptionRoomDataSource.kt
@@ -15,10 +15,10 @@ interface ExceptionRoomDataSource : BaseDao<ExceptionEntity> {
     fun findAll(): List<ExceptionEntity>
 
     @Query("SELECT * FROM ExceptionEntity WHERE id = :id")
-    suspend fun findById(id: Int): ExceptionEntity
+    suspend fun findById(id: Long): ExceptionEntity
 
     @Query("DELETE FROM ExceptionEntity WHERE id = :id")
-    suspend fun deleteById(id: Int): Int
+    suspend fun deleteById(id: Long): Int
 
     @Query("DELETE FROM ExceptionEntity")
     suspend fun deleteAll(): Int
